auth

+++ src/main/java/com/tencent/wxcloudrun/model/auth/Student.java
@@ -2,6 +2,7 @@ package com.tencent.wxcloudrun.model.auth;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 @Data
@@ -12,6 +13,6 @@ public class Student {
     public String name;
     public String college;
     public String wechatOpenid;
-    public double theta;
+    public BigDecimal theta;
     public OffsetDateTime boundAt;
 }
